ユーザ情報編集機能

+++ app/Http/Controllers/Auth/UserController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Auth;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use App\Http\Requests\EditUser;
 use Illuminate\Support\Facades\Auth;
 use App\User;
 
@@ -14,14 +15,34 @@ class UserController extends Controller
         $this->middleware('auth');
     }
 
-    public function showUserForm()
+    public function index()
     {
         return view('auth/user');
     }
 
-    public function user()
+    public function showEditForm()
     {
-        User::find(Auth::id())->delete();
-        return redirect('/login');
+        $auth = Auth::user();
+        return view('auth/edit', ['auth' => $auth]);
     }
+
+    public function edit(EditUser $request)
+    {
+        $user = Auth::user();
+
+        // 「確定」ボタン押下時の処理
+        if ($request->has("send")){
+            // 入力した値に書き換え
+            $user->name = $request->name;
+            $user->email = $request->email;
+
+            //入力した値に更新
+            $user->save();
+        }
+        else if ($request->has("delete")){
+            User::find(Auth::id())->delete();
+        }
+        return redirect()->route('users');
+    }
+
 }
